backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
transformers>=4.30.0
torch>=2.0.0
datasets>=2.10.0
sentence-transformers>=2.2.0
sentencepiece>=0.1.99
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pandas>=2.0.0
pymupdf>=1.23.0
python-docx==1.1.0
pillow>=10.0.0
pytesseract>=0.3.10
nltk>=3.8.0
spacy>=3.5.0
scikit-learn>=1.3.0